Initial code commit
Library currently has a RdcManPluginFix class with functions which
should be called in the RdcMan.IPlugin implementation.

For ease of use there is a PluginPix class which can be inherited
instead of the RdcMan.IPlugin interface which will link into RdcMan and fix
all plugin functionality. The user is then subscribe to use the events
in the class instead of using the RdcMan.IPlugin functions.